+++ plugins/fancynpcs-v2/CHANGELOG.md
@@ -1,1 +1,2 @@
 * Removed the `player_command_as_op` action, please use `player_command` and give the necessary permissions to the player instead
+* Added `turn_to_player_tracker_interval` config option